restのportに関するクラスを追加&テスト

+++ nortis-backend/src/main/java/org/nortis/application/tenant/TenantApplicationService.java
@@ -8,10 +8,13 @@ import org.nortis.domain.tenant.Tenant;
 import org.nortis.domain.tenant.TenantDomainService;
 import org.nortis.domain.tenant.TenantRepository;
 import org.nortis.domain.tenant.value.TenantId;
+import org.nortis.domain.user.AuthorityCheckDomainService;
+import org.nortis.domain.user.value.UserId;
 import org.nortis.infrastructure.annotation.ApplicationService;
 import org.nortis.infrastructure.application.ApplicationTranslator;
 import org.nortis.infrastructure.exception.DomainException;
 import org.nortis.infrastructure.message.MessageCodes;
+import org.nortis.infrastructure.security.user.NortisUserDetails;
 
 /**
  * テナントのアプリケーションサービスです
@@ -30,6 +33,34 @@ public class TenantApplicationService {
 	
 	private final AuthenticationDomainService authenticationDomainService;
 	
+	/** 権限チェックのドメインサービス */
+	private final AuthorityCheckDomainService authorityCheckDomainService;
+	
+	/**
+	 * テナントを取得します
+	 * @param <R> 結果クラス
+	 * @param rawTenantId テナントIDの文字列
+	 * @param user ユーザ
+	 * @param translator トランスレータ
+	 * @return テナント
+	 * @throws DomainException ドメインロジックエラー
+	 */
+	public <R> R getTenant(String rawTenantId, NortisUserDetails user, ApplicationTranslator<Tenant, R> translator) throws DomainException {
+		TenantId tenantId = TenantId.create(rawTenantId);
+		
+		Optional<Tenant> optTenant = this.tenantRepository.get(tenantId);
+		if (optTenant.isEmpty()) {
+			throw new DomainException(MessageCodes.nortis10003());
+		}
+
+		Tenant tenant = optTenant.get();
+
+		//権限チェック
+		authorityCheckDomainService.checkTenantAuthority(user, tenant);
+		
+		return translator.translate(tenant);
+	}
+	
 	/**
 	 * テナントを登録します
 	 * @param <R> 結果クラス
@@ -43,8 +74,12 @@ public class TenantApplicationService {
 			ApplicationTranslator<Tenant, R> translator) throws DomainException {
 		
 		TenantId tenantId = TenantId.create(command.tenantId());
+		
+		//権限のチェック
+		this.authorityCheckDomainService.checkAdminOf(UserId.create(command.user().getUsername()));
+		
 		Tenant tenant = this.tenantDomainService.createTenant(
-				tenantId, command.name(), command.userId());
+				tenantId, command.name(), command.user().getUsername());
 		
 		this.tenantRepository.save(tenant);
 		
@@ -54,10 +89,11 @@ public class TenantApplicationService {
 	/**
 	 * テナントのAPIキーを作成します
 	 * @param rawTenantId テナントID
+	 * @param user ユーザ
 	 * @return APIキー
 	 * @throws DomainException ドメインロジックエラー
 	 */
-	public ApiKey createApiKey(String rawTenantId) throws DomainException {
+	public ApiKey createApiKey(String rawTenantId, NortisUserDetails user) throws DomainException {
 		TenantId tenantId = TenantId.create(rawTenantId);
 		
 		Optional<Tenant> optTenant = this.tenantRepository.get(tenantId);
@@ -65,7 +101,11 @@ public class TenantApplicationService {
 			throw new DomainException(MessageCodes.nortis10003());
 		}
 		
-		return this.authenticationDomainService.createApiKeyOf(optTenant.get());
+		Tenant tenant = optTenant.get();
+		//権限チェック
+		authorityCheckDomainService.checkTenantAuthority(user, tenant);
+		
+		return this.authenticationDomainService.createApiKeyOf(tenant);
 	}
 	
 	/**
@@ -89,7 +129,10 @@ public class TenantApplicationService {
 		
 		Tenant tenant = optTenant.get();
 		
-		tenant.changeTenantName(command.name(), command.userId());
+		//権限チェック
+		this.authorityCheckDomainService.checkTenantAuthority(command.user(), tenant);
+		
+		tenant.changeTenantName(command.name(), command.user().getUsername());
 		
 		this.tenantRepository.update(tenant);
 		
@@ -99,10 +142,10 @@ public class TenantApplicationService {
 	/**
 	 * テナントを削除します
 	 * @param rawTenantId テナントID
-	 * @param userId ユーザID
+	 * @param user ユーザ
 	 * @throws DomainException ドメインロジックエラー
 	 */
-	public void delete(String rawTenantId, String userId) throws DomainException {
+	public void delete(String rawTenantId, NortisUserDetails user) throws DomainException {
 		TenantId tenantId = TenantId.create(rawTenantId);
 		
 		Optional<Tenant> optTenant = this.tenantRepository.get(tenantId);
@@ -110,7 +153,10 @@ public class TenantApplicationService {
 			throw new DomainException(MessageCodes.nortis10003());
 		}
 		Tenant tenant = optTenant.get();
-		tenant.deleted(userId);
+		
+		this.authorityCheckDomainService.checkTenantAuthority(user, tenant);
+		
+		tenant.deleted(user.getUsername());
 		this.tenantRepository.remove(tenant);
 	}
 }
